prepare to move bugfix patches out
